# api/templates/openapi.json
{
   "openapi": "3.1.0",
   "info": {
      "title": "biodata-registry-service",
      "description": "\n## biodata-registry-service\n\nService to fetch data from Registry.\n\n",
      "version": "0.0.0"
   },
   "paths": {
...
      "/space_admin": {
         "post": {
            "tags": [
               "admin"
            ],
            "summary": "Create Space Admin",
            "operationId": "create_space_admin",
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/SpaceAdminCreate"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/SpaceAdmins"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "get": {
            "tags": [
               "admin"
            ],
            "summary": "Get Space Admin",
            "operationId": "get_space_admin",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/SpaceAdmins"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "delete": {
            "tags": [
               "admin"
            ],
            "summary": "Delete",
            "operationId": "delete_space_admin",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {}
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "put": {
            "tags": [
               "admin"
            ],
            "summary": "Update",
            "operationId": "update_space_admin",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/SpaceAdminUpdate"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/SpaceAdmins"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/space_admins": {
         "get": {
            "tags": [
               "admin"
            ],
            "summary": "Get Space Admins",
            "operationId": "get_space_admins",
            "parameters": [
               {
                  "name": "offset",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "integer",
                     "default": 0,
                     "title": "Offset"
                  }
               },
               {
                  "name": "limit",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "integer",
                     "maximum": 1000,
                     "default": 10,
                     "title": "Limit"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "array",
                           "items": {
                              "$ref": "#/components/schemas/SpaceAdmins"
                           },
                           "title": "Response Get Space Admins"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/space": {
         "post": {
            "tags": [
               "admin"
            ],
            "summary": "Create Space",
            "operationId": "create_space",
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/SpaceCreate"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Spaces"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "get": {
            "tags": [
               "admin"
            ],
            "summary": "Get Space",
            "operationId": "get_space",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Spaces"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "delete": {
            "tags": [
               "admin"
            ],
            "summary": "Delete",
            "operationId": "delete_space",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {}
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "put": {
            "tags": [
               "admin"
            ],
            "summary": "Update",
            "operationId": "update_space",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/SpaceUpdate"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Spaces"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/spaces": {
         "get": {
            "tags": [
               "admin"
            ],
            "summary": "Get Spaces",
            "operationId": "get_spaces",
            "parameters": [
               {
                  "name": "offset",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "integer",
                     "default": 0,
                     "title": "Offset"
                  }
               },
               {
                  "name": "limit",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "integer",
                     "maximum": 1000,
                     "default": 10,
                     "title": "Limit"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "array",
                           "items": {
                              "$ref": "#/components/schemas/Spaces"
                           },
                           "title": "Response Get Spaces"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
...
   },
   "components": {
      "schemas": {
...
         "HTTPValidationError": {
            "properties": {
               "detail": {
                  "items": {
                     "$ref": "#/components/schemas/ValidationError"
                  },
                  "type": "array",
                  "title": "Detail"
               }
            },
            "type": "object",
            "title": "HTTPValidationError"
         },
...
         "SpaceAdminCreate": {
            "properties": {
               "user_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "User Id"
               },
               "space_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Space Id"
               }
            },
            "type": "object",
            "title": "SpaceAdminCreate"
         },
         "SpaceAdminUpdate": {
            "properties": {
               "user_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "User Id"
               },
               "space_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Space Id"
               }
            },
            "type": "object",
            "title": "SpaceAdminUpdate"
         },
         "SpaceAdmins": {
            "properties": {
               "user_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "User Id"
               },
               "space_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Space Id"
               },
               "id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Id"
               }
            },
            "type": "object",
            "title": "SpaceAdmins"
         },
         "SpaceCreate": {
            "properties": {
               "name": {
                  "type": "string",
                  "maxLength": 254,
                  "title": "Name"
               },
               "organization_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Organization Id"
               }
            },
            "type": "object",
            "required": [
               "name"
            ],
            "title": "SpaceCreate"
         },
         "SpaceUpdate": {
            "properties": {
               "name": {
                  "anyOf": [
                     {
                        "type": "string",
                        "maxLength": 254
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Name"
               },
               "organization_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Organization Id"
               }
            },
            "type": "object",
            "title": "SpaceUpdate"
         },
         "Spaces": {
            "properties": {
               "name": {
                  "type": "string",
                  "maxLength": 254,
                  "title": "Name"
               },
               "organization_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Organization Id"
               },
               "id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Id"
               }
            },
            "type": "object",
            "required": [
               "name"
            ],
            "title": "Spaces"
         },
...
         "ValidationError": {
            "properties": {
               "loc": {
                  "items": {
                     "anyOf": [
                        {
                           "type": "string"
                        },
                        {
                           "type": "integer"
                        }
                     ]
                  },
                  "type": "array",
                  "title": "Location"
               },
               "msg": {
                  "type": "string",
                  "title": "Message"
               },
               "type": {
                  "type": "string",
                  "title": "Error Type"
               },
               "input": {
                  "title": "Input"
               },
               "ctx": {
                  "type": "object",
                  "title": "Context"
               }
            },
            "type": "object",
            "required": [
               "loc",
               "msg",
               "type"
            ],
            "title": "ValidationError"
         }
      }
   }
}